Clean up discretization types
Now that the `discretization_info` abstract type is defined, can define
the specific types for each implementation in the module along with the
implementation. This makes more sense and simplifies the dependencies
between different modules.
