+++ src/moment_kinetics_structs.jl
@@ -4,7 +4,6 @@ cycles when they are used by several other modules.
 """
 module moment_kinetics_structs
 
-using FFTW
 using ..communication
 using ..type_definitions: mk_float
 
@@ -51,31 +50,6 @@ discretization_info for one dimension
 All the specific discretizations in moment_kinetics are subtypes of this type.
 """
 abstract type discretization_info end
-
-"""
-Chebyshev pseudospectral discretization
-"""
-struct chebyshev_info{TForward <: FFTW.cFFTWPlan, TBackward <: AbstractFFTs.ScaledPlan} <: discretization_info
-    # fext is an array for storing f(z) on the extended domain needed
-    # to perform complex-to-complex FFT using the fact that f(theta) is even in theta
-    fext::Array{Complex{mk_float},1}
-    # Chebyshev spectral coefficients of distribution function f
-    # first dimension contains location within element
-    # second dimension indicates the element
-    f::Array{mk_float,2}
-    # Chebyshev spectral coefficients of derivative of f
-    df::Array{mk_float,1}
-    # plan for the complex-to-complex, in-place, forward Fourier transform on Chebyshev-Gauss-Lobatto grid
-    forward::TForward
-    # plan for the complex-to-complex, in-place, backward Fourier transform on Chebyshev-Gauss-Lobatto grid
-    #backward_transform::FFTW.cFFTWPlan
-    backward::TBackward
-end
-
-"""
-Finite difference discretization
-"""
-struct finite_difference_info <: discretization_info end
 
 """
 Type representing a spatial dimension with only one grid point
